+++ packages/lint/src/index.ts
@@ -2,6 +2,7 @@ import buttonRules from './rules/button-rules.js';
 import containerRules from './rules/container-rules.js';
 import imageRules from './rules/image-rules.js';
 import linkRules from './rules/link-rules.js';
+import proseRules from './rules/prose-rules.js';
 import textRules from './rules/text-rules.js';
 
 const rules = {
@@ -9,6 +10,7 @@ const rules = {
   ...containerRules.rules,
   ...imageRules.rules,
   ...linkRules.rules,
+  ...proseRules.rules,
   ...textRules.rules,
 };
 
@@ -19,6 +21,7 @@ const configs = {
       ...containerRules.configs.recommended.rules,
       ...imageRules.configs.recommended.rules,
       ...linkRules.configs.recommended.rules,
+      ...proseRules.configs.recommended.rules,
       ...textRules.configs.recommended.rules,
     },
   },
@@ -40,6 +43,8 @@ export type {
   ImageTagAnalysis,
   LinkClassInfo,
   LinkTagAnalysis,
+  ProseClassInfo,
+  ProseTagAnalysis,
   TextComponentInfo,
   TextTagAnalysis,
 } from './types.js';