+++ sandbox/vue-library/src/index.ts
@@ -2,6 +2,7 @@ import CedarButtonFixture from './components/CedarButtonFixture.vue';
 import CedarContainerFixture from './components/CedarContainerFixture.vue';
 import CedarImageFixture from './components/CedarImageFixture.vue';
 import CedarLinkFixture from './components/CedarLinkFixture.vue';
+import CedarProseFixture from './components/CedarProseFixture.vue';
 import CedarTextFixture from './components/CedarTextFixture.vue';
 import IconUser from './components/IconUser.vue';
 
@@ -10,6 +11,7 @@ export {
   CedarContainerFixture,
   CedarImageFixture,
   CedarLinkFixture,
+  CedarProseFixture,
   CedarTextFixture,
   IconUser,
 };